Move module_id functions to separate module
